% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EML2df.R
\name{EML2df}
\alias{EML2df}
\title{Convert a corpus of EML documents into tables}
\usage{
EML2df(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(list) List of data.frames each corresponding to output from a get_ function
}
\description{
Convert a corpus of EML documents into tables
}
\details{
This is a simple wrapper around the get_ functions also exported from this package.
}
