% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotations}
\alias{get_annotations}
\title{Title}
\usage{
get_annotations(corpus)
}
\arguments{
\item{corpus}{(list) list of EML documents, output from import_corpus}
}
\value{
(data.frame) Table of annotation metadata from dataset, entity, and/or attribute
}
\description{
Title
}
