% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change_history.R
\name{get_changehistory}
\alias{get_changehistory}
\title{Title}
\usage{
get_changehistory(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(data.frame) Table of change history metadata
}
\description{
Title
}
