% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corpus.R
\name{get_corpus}
\alias{get_corpus}
\title{Download EML metadata from all data packages in an Environmental Data Initiative (EDI) scope.}
\usage{
get_corpus(scope, path, EDI_env = "production")
}
\arguments{
\item{scope}{(character) EDI scope, e.g. "knb-lter-ble"}

\item{path}{(character) Path to existing directory to download EML files into}

\item{EDI_env}{(character) EDI server to query. Defaults to "production". Options are "production", "staging", or "development".}
}
\value{
Nothing in R, just download files.
}
\description{
This function fetchs newest IDs and revision numbers from a given EDI scope, then download EML metadata files to a specified directory. Newest revision from each package only. Files will be named after the full packageId as found in the metadata, e.g. "knb-lter-ble.1.7.xml".
}
