% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage_geo.R
\name{get_coverage_geo}
\alias{get_coverage_geo}
\title{Get geographicCoverage}
\usage{
get_coverage_geo(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(data.frame) Table of geographic coverage metadata
}
\description{
Get geographicCoverage
}
