% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage_tax.R
\name{get_coverage_tax}
\alias{get_coverage_tax}
\title{Get taxonomic coverage}
\usage{
get_coverage_tax(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(data.frame) table of taxonomic coverage metadata. Since taxonomicClassification can be recursively nested, each taxonomicClassification node gets its own row.
}
\description{
Get taxonomic coverage
}
