% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get dataset level metadata from EML corpus in table form}
\usage{
get_dataset(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents in EMLD format, each named according to the package ID listed in EML, e.g. "knb-lter-mcr.1.1". Output from import_corpus.}
}
\value{
data.frame
}
\description{
Get dataset level metadata from EML corpus in table form
}
