% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasetlevel_element.R
\name{get_datasetlevel_element}
\alias{get_datasetlevel_element}
\title{Get custom dataset-level metadata}
\usage{
get_datasetlevel_element(corpus, element_names, parse_function)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}

\item{element_names}{(character) A string specifying the dataset-level. If the desired node is not an immediate child node of dataset, than supply a character vector to specify the xpath leading to the desired node. For example, to get geographicCoverage, say c("coverage", "geographicCoverage").}

\item{parse_function}{(function) A parse function, taking a EML node and outputting a data.frame}
}
\value{
(data.frame) A table containing parsed metadata from the specified node according to the supplied parse function.
}
\description{
Get custom dataset-level metadata
}
\details{
this function will loop through all datasets in the supplied corpus and output a table of parsed metadata from the specified node according to the supplied parse function. Note that dataset-level here is meant different from dataset metadata that get_dataset() gets. Dataset-level refers to nodes that have child nodes in themselves, such as maintenance, but appear under dataset in a EML document.
}
