% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entities.R
\name{get_entities}
\alias{get_entities}
\title{Get information about entities}
\usage{
get_entities(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(data.frame) Parsed entity metadata
}
\description{
Get information about entities
}
\details{
This function will loop through all dataTable, spatialVector, spatialRaster, otherEntity, and view nodes across datasets in the corpus and output a table with info about the entity itself: name, description, numberofRecords, objectName (aka file name), file size, hash value.
}
