% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keywords.R
\name{get_keywords}
\alias{get_keywords}
\title{Get keywords}
\usage{
get_keywords(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(data.frame) parsed keywords
}
\description{
Get keywords
}
