% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_methods}
\alias{get_methods}
\title{Get methods}
\usage{
get_methods(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(data.frame) Table of methods
}
\description{
Get methods
}
