% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multilevel_element.R
\name{get_multilevel_element}
\alias{get_multilevel_element}
\title{Title}
\usage{
get_multilevel_element(corpus, element_names, parse_function)
}
\arguments{
\item{parse_function}{}
}
\value{

}
\description{
Title
}
