% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parties.R
\name{get_parties}
\alias{get_parties}
\title{Get personnel}
\usage{
get_parties(corpus)
}
\arguments{
\item{corpus}{(list) List of EML documents, output from import_corpus}
}
\value{
(data.frame) Table of personnel information: name, address, organization, role, position, email etc
}
\description{
Get personnel
}
\details{
This function will grab responsibleParty-type nodes from these nodes: creator, associatedParty, contact, publisher, metadataProvider, and project/personnel, and output a table of personnel information.
}
