% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_multiple}
\alias{handle_multiple}
\title{handles instances where there are multiple in a list and collapses them int a string}
\usage{
handle_multiple(x, paste = T)
}
\description{
handles instances where there are multiple in a list and collapses them int a string
}
