% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{normalize_tables}
\alias{normalize_tables}
\title{Title}
\usage{
normalize_tables(tables, clean = FALSE)
}
\arguments{
\item{clean}{(logical) TRUE/FALSE on whether pkEML should attempt to clean the input tables before normalizing. Cleaning entails operations like trimming whitespace, case-standardization, removing non-alphanumeric symbols. pkEML will attempt to preserve the exact content extracted from EML where it matters, i.e. enumeration}
}
\value{

}
\description{
Title
}
