% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{parse_annotation}
\alias{parse_annotation}
\title{Title}
\usage{
parse_annotation(x)
}
\arguments{
\item{x}{(list) single annotation node}
}
\value{
(data.frame) table of parsed annotation
}
\description{
Title
}
