% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change_history.R
\name{parse_changehistory}
\alias{parse_changehistory}
\title{Title}
\usage{
parse_changehistory(x)
}
\arguments{
\item{x}{(list) changeHistory node}
}
\value{
(data.frame) parsed change history
}
\description{
Title
}
