% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_generic}
\alias{parse_generic}
\title{A generic parse function}
\usage{
parse_generic(x)
}
\arguments{
\item{x}{(list) node to parse}
}
\value{
(data.frame) Parsed node
}
\description{
A generic parse function
}
