% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keywords.R
\name{parse_keyword}
\alias{parse_keyword}
\title{Parse keyword}
\usage{
parse_keyword(x)
}
\arguments{
\item{x}{(list) single keyword node}
}
\value{
(data.frame) parsed node
}
\description{
Parse keyword
}
