% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keywords.R
\name{parse_keywordset}
\alias{parse_keywordset}
\title{Parse keywordSet}
\usage{
parse_keywordset(x)
}
\arguments{
\item{x}{(list) single keywordSet node}
}
\value{
(data.frame) parsed node
}
\description{
Parse keywordSet
}
