% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{parse_methodstep}
\alias{parse_methodstep}
\title{Parse methodStep}
\usage{
parse_methodstep(x)
}
\arguments{
\item{x}{(list) single methodStep node}
}
\value{
(data.frame) parsed node
}
\description{
Parse methodStep
}
