% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_packageId}
\alias{parse_packageId}
\title{Parse packageId}
\usage{
parse_packageId(full_id)
}
\arguments{
\item{full_id}{(character) Package ID in EML style, e.g. "knb-lter-mcr.1.1"}
}
\value{
List of three items: "scope", "id", "rev".
}
\description{
Take the packageId field in EML and break it down into scope, id, and revision number. If packageId does not conform to the scope, id, revision number pattern, the function will just return the whole ID in each field.
}
