% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parties.R
\name{parse_party}
\alias{parse_party}
\title{parse responsible party}
\usage{
parse_party(x)
}
\arguments{
\item{x}{(list) a responsible party node}
}
\value{
(data.frame) parsed node
}
\description{
parse responsible party
}
