% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{parse_project}
\alias{parse_project}
\title{parse project node}
\usage{
parse_project(x)
}
\arguments{
\item{x}{(list) single project node}
}
\value{
(data.frame) parsed node
}
\description{
parse project node
}
