% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recursive_check}
\alias{recursive_check}
\title{Check if chain of descending element names is present in node}
\usage{
recursive_check(x, element_names)
}
\arguments{
\item{x}{(list) node to check}

\item{element_names}{(character) Name or vector of descending names to check}
}
\value{
(logical) TRUE if chain of element names are present in the node in that order, FALSE if not.
}
\description{
Check if chain of descending element names is present in node
}
