% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_reference}
\alias{resolve_reference}
\title{resolve reference in a node}
\usage{
resolve_reference(x, element_name, eml)
}
\arguments{
\item{x}{(list) node of element}

\item{element_name}{(character) name of element}

\item{eml}{(list) full EML document}
}
\value{
(list) If the node contains a reference to some other element in the document, the function returns the same node, just with the reference replaced by the contents of the referenced node. If not, this returns the exact same node and does nothing to it.
}
\description{
resolve reference in a node
}
