% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_reference_all}
\alias{resolve_reference_all}
\title{Resolve all references in a EML document}
\usage{
resolve_reference_all(eml)
}
\arguments{
\item{eml}{full EML document}
}
\value{
(list) full EML document with all references resolved and replaced with each respective referenced node
}
\description{
Resolve all references in a EML document
}
